Fix links for Jenkins and GitHub issues.
Also updating branch update link. Main is no longer the default branch for PRs. Development or stabalization should be targeted based on the change.

Nightly pipeline is only running on the jenkins-pipeline instance which is not accessible publicly. Link will be re-added once it's enabled in jenkins.build.o3de.org.

Signed-off-by: brianherrera <briher@amazon.com>